FIX: Split all code in different files
